Corrected question #14

+++ src/main/java/com/revature/eval/java/core/EvaluationService.java
@@ -510,7 +510,7 @@ public class EvaluationService {
 			throw new IllegalArgumentException();
 		}
 		try {
-			Integer.parseInt(string);
+			Long.parseLong(string);
 		} catch (NumberFormatException nfe) {
 			throw new IllegalArgumentException();
 		}
